requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
cnbc==0.0.2
dj-database-url==2.1.0
Django==4.2.7
django-cors-headers==4.3.0
django-extensions==3.2.3
django-ninja==0.22.2
django-on-heroku==1.1.2
idna==3.4
install==1.3.5
psycopg2-binary==2.9.9
pydantic==1.10.13
python-decouple==3.8
python-dotenv==1.0.0
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.7
whitenoise==6.6.0
